Streamlit/requirements.txt
streamlit
pandas
numpy
paho-mqtt
mysql-connector-python
firebase-admin
plotly
